Multilevel functionality added
Added the multilevel/spin.py. Turns out to be mostly required for even a single state since we need to be able to differentiate a levelled state from a spin state.

The custom spin file adds AtomKets, which have largely the same usage as Spin kets in SymPy, but lack coupling with one another.

TODO: AtomKets are not properly represented yet. Try to fix this.