added a way to sub in variables
Currently a very difficult way of doing it as it requires knowledge of every label that is used. This isn't such a problem when we define the labels ourselves, but can become an issue when relying on automatic label generation.

Either we add a method that makes it clear what labels have been used, or force the user to give their own labels.